pyproject.toml
[project]
name = "aws-hackathon-20250915"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore>=0.1.3",
    "bedrock-agentcore-starter-toolkit>=0.1.10",
    "strands-agents>=1.8.0",
    "strands-agents-tools>=0.2.7",
    "boto3>=1.26.0",
]
